pyproject.toml
[project]
name = "eurosatproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.12"
dependencies = [
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.5.0",
    "numpy<2",
    "scikit-learn>=1.0.0",
    "seaborn>=0.11.0",
    "torch==1.13.1+cu117",
    "torchvision==0.14.1+cu117",
    "tqdm>=4.60.0",
    "Pillow>=9.0.0",
    "earthengine-api>=0.1.0",
    "geemap>=0.20.0",
    "dotenv>=0.9.9",
    "rasterio>=1.4.4",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu117" }]
torchvision = [{ index = "pytorch-cu117" }]

[[tool.uv.index]]
name = "pytorch-cu117"
url = "https://download.pytorch.org/whl/cu117"
explicit = true
